backend/ems/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.1
Django==5.1.6
djangorestframework==3.15.2
pycparser==2.22
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
